backend/requirements.txt
# requirements.txt
flask==2.0.1
requests==2.26.0
beautifulsoup4==4.9.3
google-cloud-pubsub==2.8.0
python-dotenv==0.19.0
reportlab==3.6.2